refactor and move information schema related method to information schema
